Make chart component generic.
